break out NoProjectSerializer
