+++ api/serializers/__init__.py
@@ -32,3 +32,4 @@ from .provider_machine_serializer import ProviderMachineSerializer
 from .paginated_provider_machine_serializer import PaginatedProviderMachineSerializer
 from .group_serializer import GroupSerializer
 from .volume_serializer import VolumeSerializer
+from .no_project_serializer import NoProjectSerializer
